Implement Simple Tabulation
In spite of its simple implementation, Simple Tabulation provides many desired
guarantees, such as linear probing. Reference:
    Thorup, Mikkel. "Fast and powerful hashing using tabulation."
    Communications of the ACM 60.7 (2017): 94-101.
    https://dl.acm.org/doi/abs/10.1145/3068772

This is just a proof of concept to figure out if Simple Tabulation could
improve the performance of the current Ruby Hashing implementation. The code is
not in the correct place and it is only implemented for 64 bits and for the
general case (excluding strings, numbers and symbols). But this is enough to
prove the potential of Simple Tabulation. Below are the results got executing
the following code:
https://gist.github.com/Ana06/cc6f3a737ad16548996a22718de8b01b

Time to add 600000 objects to an empty hash 100 times:
    - master (without Simple Tabulation): 29.68
    - master with Linear Probing (without Simple Tabulation): 99.76
    - master with Quadratic Probing (without Simple Tabulation): 29.97
    - master with Simple Tabulation: 15.76
    - master with Linear Probing and Simple Tabulation: 24.23
    - master with Quadratic Probing and Simple Tabulation: 13.92

The results are in seconds. `master` refers to ruby 2.8.0dev
(2020-05-07T16:22:38Z master 7ded8fd29a) [x86_64-linux].
I tried with 8 x Intel i5-8265U at 1.60 GHz.