Enable Quadratic Probing
Enable the already implemented Quadratic Probing algorithm. It was introduced
in the following commit, alghough not enabled (it was not even compiled):
75775157ea68efdd0b71345a37a6930e5bf1d2ef

If we call `h(x)` to the hash function which maps `x` to a position in the
hash, then the i-th probe position for `x` is given by the following function:
`h(x,i) = h(x) + 1/2i + 1/2i^2`

This produces the following probe sequence:
`h(x), h(x) + 1, h(x) + 3, h(x) + 6, h(x) + 10, ...`
In this sequence the increase values are triangular numbers, making it
efficient to implement.

Quadratic Probing may perform better than the current enabled implementation
with Tabulation Hashing, which I plan to implement next.